refactor(backend): dedupe GitHub URL extraction
